Refactored backend, added remember me functionality

+++ workout-tracker-backend/src/main/java/com/jok92/workout_tracker_backend/services/auth/RefreshService.java
@@ -2,6 +2,7 @@ package com.jok92.workout_tracker_backend.services.auth;
 
 import com.jok92.workout_tracker_backend.models.auth.AccessRefreshPair;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
@@ -16,6 +17,7 @@ public class RefreshService {
 
     @Autowired
     private RedisService redisService;
+
 
     public AccessRefreshPair issueNewAccessRefreshPair(UUID userId) {
         return new AccessRefreshPair(jwtService.generateToken(userId), generateRefreshToken(userId));
@@ -33,7 +35,7 @@ public class RefreshService {
 
     public UUID generateRefreshToken(UUID userID) {
         UUID token = UUID.randomUUID();
-        long timeout = 7;
+        int timeout = 7;
         TimeUnit unit = TimeUnit.DAYS;
         redisService.saveWithExpiry(token.toString(), userID.toString(), timeout, unit);
         return token;
